feat(Presentation): :sparkles: loation variable registry
location variable for latitude and longitude (permission)

It could be called from the sign_up_screen the location regsitry
